# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDagRunReply": {
        "properties": {
          "acceptedByThreshold": {
            "default": true,
            "title": "Acceptedbythreshold",
            "type": "boolean"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AgentDagAttempt"
            },
            "title": "Attempts",
            "type": "array"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "levels": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Levels",
            "type": "array"
          },
          "synthesis": {
            "$ref": "#/components/schemas/AgentRunReply"
          },
          "taskResults": {
            "items": {
              "$ref": "#/components/schemas/AgentDagTaskResult"
            },
            "title": "Taskresults",
            "type": "array"
          },
          "tenantId": {
            "title": "Tenantid",
            "type": "string"
          }
        },
        "required": [
          "goal",
          "synthesis",
          "tenantId"
        ],
        "title": "AgentDagRunReply",
        "type": "object"
      },
...
      "AgentPlanRunRequest": {
        "properties": {
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "webhookUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhookurl"
          }
        },
        "title": "AgentPlanRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Incremental AgentScope 2.0 replacement for langchain4j-platform agent-service.",
    "title": "AgentScope Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agent/analyst/run": {
      "post": {
        "operationId": "run_analyst_agent_agent_analyst_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDagRunReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Analyst Agent",
        "tags": [
          "agent-analyst"
        ]
      }
    },
    "/agent/analyst/run/async": {
      "post": {
        "operationId": "run_analyst_agent_async_agent_analyst_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAsyncTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Analyst Agent Async",
        "tags": [
          "agent-analyst"
        ]
      }
    },
...
    "/agent/chain": {
      "post": {
        "operationId": "run_prompt_chain_agent_chain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ChainRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainRunReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Prompt Chain",
        "tags": [
          "agent-chain"
        ]
      }
    },
    "/agent/dag/plan-run": {
      "post": {
        "operationId": "plan_and_run_agent_dag_agent_dag_plan_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDagRunReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan And Run Agent Dag",
        "tags": [
          "agent-dag"
        ]
      }
    },
    "/agent/dag/plan-run/async": {
      "post": {
        "operationId": "plan_and_run_agent_dag_async_agent_dag_plan_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAsyncTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan And Run Agent Dag Async",
        "tags": [
          "agent-dag"
        ]
      }
    },
    "/agent/dag/run": {
      "post": {
        "operationId": "run_agent_dag_agent_dag_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDagRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDagRunReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Agent Dag",
        "tags": [
          "agent-dag"
        ]
      }
    },
    "/agent/dag/run/async": {
      "post": {
        "operationId": "run_agent_dag_async_agent_dag_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDagRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAsyncTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Agent Dag Async",
        "tags": [
          "agent-dag"
        ]
      }
    },
    "/agent/process/run": {
      "post": {
        "description": "Queries workflow status, pending tasks, or policy only. This candidate never starts, approves, claims, or modifies a workflow.",
        "operationId": "run_readonly_process_agent_process_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDagRunReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the read-only Process candidate",
        "tags": [
          "agent-process"
        ]
      }
    },
    "/agent/process/run/async": {
      "post": {
        "operationId": "run_readonly_process_async_agent_process_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentPlanRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAsyncTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Readonly Process Async",
        "tags": [
          "agent-process"
        ]
      }
    },
    "/agent/reflexive": {
      "post": {
        "operationId": "run_reflexion_agent_reflexive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReflexionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflexionReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Reflexion",
        "tags": [
          "agent-reflexion"
        ]
      }
    },
    "/agent/reflexive/stream": {
      "post": {
        "operationId": "stream_reflexion_agent_reflexive_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReflexionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Reflexion",
        "tags": [
          "agent-reflexion"
        ]
      }
    },
...
  }
}